requirements.txt
torchmetrics
pytorch-lightning
ftfy